requirements.txt
streamlit
newspaper3k
transformers
torch
langdetect
